Create Facebook_Hacker_Cup_2016_Lazy_Loading
C solution for the Lazy Loading Problem which was asked in the Facebook Hacker Cup 2016 Qualification Round.
Lazy Loading can be solved in a Greedy Approach, which would take O(nlogn) time in the worst case. However, I have made improvements which would also make it work in O(n) time in the best case.